+++ wayfinder_paths/adapters/hyperliquid_adapter/adapter.py
@@ -31,7 +31,7 @@ from wayfinder_paths.adapters.hyperliquid_adapter.utils import spot_index_from_a
 from wayfinder_paths.core.adapters.BaseAdapter import BaseAdapter
 from wayfinder_paths.core.clients.HyperliquidInfoClient import HYPERLIQUID_INFO_CLIENT
 from wayfinder_paths.core.clients.HyperliquidQuicknodeInfoClient import (
-    HYPERLIQUID_QN_CLIENT,
+    HYPERLIQUID_QUICKNODE_CLIENT,
 )
 from wayfinder_paths.core.constants import ZERO_ADDRESS
 from wayfinder_paths.core.constants.contracts import (
@@ -456,7 +456,7 @@ class HyperliquidAdapter(BaseAdapter):
         self, address: str
     ) -> tuple[Literal[True], dict[str, Any]] | tuple[Literal[False], str]:
         try:
-            portfolio = await HYPERLIQUID_QN_CLIENT.portfolio_state(address)
+            portfolio = await HYPERLIQUID_QUICKNODE_CLIENT.portfolio_state(address)
         except Exception as exc:
             self.logger.error(f"Failed to fetch user_state for {address}: {exc}")
             return False, str(exc)
