+++ wayfinder_paths/core/clients/test_hyperliquid_info_client_live.py
@@ -1,13 +1,7 @@
-"""Live network tests for HyperliquidInfoClient. Hits real Hyperliquid
-public + vault-backend QN proxy — no mocks.
+"""Live network tests for HyperliquidInfoClient (public HL path).
+Hits api.hyperliquid.xyz directly — no auth needed.
 
-Set RUN_HYPERLIQUID_LIVE_TESTS=1 to enable. Tests against whatever
-`get_api_base_url()` resolves to (default: prod). Override via config.json
-or WAYFINDER_CONFIG_PATH if you want dev.
-
-QN-proxied tests require WAYFINDER_API_KEY in env (or config.json) — they're
-skipped if missing. Public-direct tests run unconditionally once the module
-is enabled.
+Set RUN_HYPERLIQUID_LIVE_TESTS=1 to enable.
 """
 
 from __future__ import annotations
@@ -18,11 +12,6 @@ import time
 import pytest
 
 from wayfinder_paths.core.clients.HyperliquidInfoClient import HyperliquidInfoClient
-from wayfinder_paths.core.clients.HyperliquidQuicknodeInfoClient import (
-    QN_PROXIED_TYPES,
-    HyperliquidQuicknodeInfoClient,
-)
-from wayfinder_paths.core.config import get_api_key
 
 if os.getenv("RUN_HYPERLIQUID_LIVE_TESTS", "").lower() not in ("1", "true", "yes"):
     pytest.skip(
@@ -30,122 +19,12 @@ if os.getenv("RUN_HYPERLIQUID_LIVE_TESTS", "").lower() not in ("1", "true", "yes
         allow_module_level=True,
     )
 
-# Active HL wallet (from QN's own docs examples) — has perp + spot state
-# across multiple dexes, so positive-data assertions are stable.
 TEST_USER = "0xf9a9a403b039996082049394935f815523157330"
-TEST_BUILDER = "0xaa1d89f333857ed78f8434cc4f896a9293efe65c"
-
-needs_api_key = pytest.mark.skipif(
-    not get_api_key(),
-    reason="WAYFINDER_API_KEY not configured — QN-proxied path requires backend auth",
-)
 
 
 @pytest.fixture
 def client() -> HyperliquidInfoClient:
-    # Function-scoped — httpx.AsyncClient must be built inside the test's
-    # event loop, not at module import.
     return HyperliquidInfoClient()
-
-
-# ── Whitelist sanity ──────────────────────────────────────────────────────
-
-
-def test_whitelist_covers_critical_methods() -> None:
-    for method in (
-        "clearinghouseState",
-        "spotClearinghouseState",
-        "frontendOpenOrders",
-        "maxBuilderFee",
-        "meta",
-        "openOrders",
-        "perpDexs",
-        "spotMeta",
-    ):
-        assert method in QN_PROXIED_TYPES
-
-
-# ── QN-proxied path (backend → QuickNode) ─────────────────────────────────
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_clearinghouse_state(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "clearinghouseState", "user": TEST_USER})
-    assert "marginSummary" in r
-    assert "assetPositions" in r
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_spot_clearinghouse_state(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "spotClearinghouseState", "user": TEST_USER})
-    assert "balances" in r
-    assert isinstance(r["balances"], list)
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_meta(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "meta"})
-    assert "universe" in r
-    assert len(r["universe"]) > 50
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_spot_meta(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "spotMeta"})
-    assert "tokens" in r and "universe" in r
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_open_orders(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "openOrders", "user": TEST_USER})
-    assert isinstance(r, list)
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_perp_dexes(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "perpDexs"})
-    assert isinstance(r, list)
-    assert any(d and d["name"] == "xyz" for d in r if d is not None)
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_max_builder_fee(client: HyperliquidInfoClient) -> None:
-    r = await client.post(
-        {"type": "maxBuilderFee", "user": TEST_USER, "builder": TEST_BUILDER}
-    )
-    assert isinstance(r, int)
-    assert r >= 0
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_qn_frontend_open_orders(client: HyperliquidInfoClient) -> None:
-    r = await client.post({"type": "frontendOpenOrders", "user": TEST_USER})
-    assert isinstance(r, list)
-
-
-# ── Portfolio state (backend → QuickNode hl_portfolioState) ───────────────
-
-
-@needs_api_key
-@pytest.mark.asyncio
-async def test_portfolio_state() -> None:
-    r = await HyperliquidQuicknodeInfoClient().portfolio_state(TEST_USER)
-    assert "clearinghouseState" in r
-    assert "spotClearinghouseState" in r
-    assert "userAbstraction" in r
-    assert "assetPositions" in r["clearinghouseState"]
-    assert "balances" in r["spotClearinghouseState"]
-
-
-# ── Public-direct path (SDK Info → api.hyperliquid.xyz) ──────────────────
 
 
 @pytest.mark.asyncio
@@ -240,9 +119,6 @@ async def test_public_user_fills_by_time(client: HyperliquidInfoClient) -> None:
     assert isinstance(r, list)
 
 
-# Adapter-call-site coverage — public-only HL types the adapter hits today.
-
-
 @pytest.mark.asyncio
 async def test_public_all_perp_metas(client: HyperliquidInfoClient) -> None:
     r = await client.post({"type": "allPerpMetas"})
@@ -266,7 +142,6 @@ async def test_public_user_abstraction(client: HyperliquidInfoClient) -> None:
 
 @pytest.mark.asyncio
 async def test_public_margin_table(client: HyperliquidInfoClient) -> None:
-    # HL margin table 1 is the default cross-margin schedule — always present.
     r = await client.post({"type": "marginTable", "id": 1})
     assert isinstance(r, dict)
     assert "marginTiers" in r or "description" in r
